requirements.txt
fastapi==0.115.14
sqlmodel==0.0.24
pymysql
python-dotenv==1.1.1
uvicorn==0.35.0
jinja2==3.1.4
cryptography